+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/event/handler/internal/utils/SelfUserUtils.kt
@@ -35,8 +35,7 @@ internal fun MutableGlobalState.updateCurrentUser(self: SelfUser) {
     }
 
     clientState.toMutableState()?.setUser(me)
-    setBanned(me.banned)
-    setBanned(me.banned)
+    setBanned(me.isBanned)
     setMutedUsers(me.mutes)
     setChannelMutes(me.channelMutes)
     setTotalUnreadCount(me.totalUnreadCount)