+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 ### ✅ Added
 
 ### ⚠️ Changed
+- 🚨Breaking change: The following properties in `User` class are now **nullable**: `banned`, `invisible`. [#5105](https://github.com/GetStream/stream-chat-android/pull/5105)
+  * Please use `User.isBanned` as non-nullable version. 
+  * Please use `User.isInvisible` as non-nullable version. 
+  * Properties such as `invisible`, `banned`, `teams` and `role` are not being used to establish WS connection flow if not specified.
 
 ### ❌ Removed